+++ tests/test_qvm.py
@@ -42,12 +42,13 @@ class TestQVMBasic(BaseTest):
 
     # pylint: disable=protected-access
 
-    def test_identity_expectation(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_identity_expectation(self, shots, qvm, compiler, parallel):
         """Test that identity expectation value (i.e. the trace) is 1"""
         theta = 0.432
         phi = 0.123
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=parallel)
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(theta, wires=[0])
@@ -65,12 +66,13 @@ class TestQVMBasic(BaseTest):
         # below are the analytic expectation values for this circuit (trace should always be 1)
         self.assertAllAlmostEqual(res, np.array([1, 1]), delta=3 / np.sqrt(shots))
 
-    def test_pauliz_expectation(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_pauliz_expectation(self, shots, qvm, compiler, parallel):
         """Test that PauliZ expectation value is correct"""
         theta = 0.432
         phi = 0.123
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=parallel)
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(theta, wires=[0])
@@ -86,15 +88,18 @@ class TestQVMBasic(BaseTest):
 
         # below are the analytic expectation values for this circuit
         self.assertAllAlmostEqual(
-            res, np.array([np.cos(theta), np.cos(theta) * np.cos(phi)]), delta=3 / np.sqrt(shots)
+            res,
+            np.array([np.cos(theta), np.cos(theta) * np.cos(phi)]),
+            delta=3 / np.sqrt(shots),
         )
 
-    def test_paulix_expectation(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_paulix_expectation(self, shots, qvm, compiler, parallel):
         """Test that PauliX expectation value is correct"""
         theta = 0.432
         phi = 0.123
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=parallel)
 
         with qml.tape.QuantumTape() as tape:
             qml.RY(theta, wires=[0])
@@ -109,15 +114,18 @@ class TestQVMBasic(BaseTest):
         res = np.array([dev.expval(O1.obs), dev.expval(O2.obs)])
         # below are the analytic expectation values for this circuit
         self.assertAllAlmostEqual(
-            res, np.array([np.sin(theta) * np.sin(phi), np.sin(phi)]), delta=3 / np.sqrt(shots)
+            res,
+            np.array([np.sin(theta) * np.sin(phi), np.sin(phi)]),
+            delta=3 / np.sqrt(shots),
         )
 
-    def test_pauliy_expectation(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_pauliy_expectation(self, shots, qvm, compiler, parallel):
         """Test that PauliY expectation value is correct"""
         theta = 0.432
         phi = 0.123
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=parallel)
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(theta, wires=[0])
@@ -136,7 +144,8 @@ class TestQVMBasic(BaseTest):
             res, np.array([0, -np.cos(theta) * np.sin(phi)]), delta=3 / np.sqrt(shots)
         )
 
-    def test_hadamard_expectation(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_hadamard_expectation(self, shots, qvm, compiler, parallel):
         """Test that Hadamard expectation value is correct"""
         theta = 0.432
         phi = 0.123
@@ -157,7 +166,10 @@ class TestQVMBasic(BaseTest):
 
         # below are the analytic expectation values for this circuit
         expected = np.array(
-            [np.sin(theta) * np.sin(phi) + np.cos(theta), np.cos(theta) * np.cos(phi) + np.sin(phi)]
+            [
+                np.sin(theta) * np.sin(phi) + np.cos(theta),
+                np.cos(theta) * np.cos(phi) + np.sin(phi),
+            ]
         ) / np.sqrt(2)
         self.assertAllAlmostEqual(res, expected, delta=3 / np.sqrt(shots))
 
@@ -171,7 +183,7 @@ class TestQVMBasic(BaseTest):
         theta = 0.432
         phi = 0.123
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=True)
 
         with qml.tape.QuantumTape() as tape:
             qml.RY(theta, wires=[0])
@@ -190,13 +202,20 @@ class TestQVMBasic(BaseTest):
         a = H[0, 0]
         re_b = H[0, 1].real
         d = H[1, 1]
-        ev1 = ((a - d) * np.cos(theta) + 2 * re_b * np.sin(theta) * np.sin(phi) + a + d) / 2
-        ev2 = ((a - d) * np.cos(theta) * np.cos(phi) + 2 * re_b * np.sin(phi) + a + d) / 2
+        ev1 = (
+            (a - d) * np.cos(theta) + 2 * re_b * np.sin(theta) * np.sin(phi) + a + d
+        ) / 2
+        ev2 = (
+            (a - d) * np.cos(theta) * np.cos(phi) + 2 * re_b * np.sin(phi) + a + d
+        ) / 2
         expected = np.array([ev1, ev2])
 
         self.assertAllAlmostEqual(res, expected, delta=4 / np.sqrt(shots))
 
-    def test_multi_qubit_hermitian_expectation(self, shots, execution_timeout, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_multi_qubit_hermitian_expectation(
+        self, shots, execution_timeout, qvm, compiler, parallel
+    ):
         """Test that arbitrary multi-qubit Hermitian expectation values are correct"""
         theta = 0.432
         phi = 0.123
@@ -210,7 +229,12 @@ class TestQVMBasic(BaseTest):
             ]
         )
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=10 * shots, execution_timeout=execution_timeout)
+        dev = plf.QVMDevice(
+            device="2q-qvm",
+            shots=10 * shots,
+            execution_timeout=execution_timeout,
+            parallel=parallel,
+        )
 
         with qml.tape.QuantumTape() as tape:
             qml.RY(theta, wires=[0])
@@ -234,9 +258,10 @@ class TestQVMBasic(BaseTest):
 
         self.assertAllAlmostEqual(res, expected, delta=5 / np.sqrt(shots))
 
-    def test_var(self, shots, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_var(self, shots, qvm, compiler, parallel):
         """Tests for variance calculation"""
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        dev = plf.QVMDevice(device="2q-qvm", shots=shots, parallel=parallel)
 
         phi = 0.543
         theta = 0.6543
@@ -250,13 +275,21 @@ class TestQVMBasic(BaseTest):
         dev._samples = dev.generate_samples()
 
         var = np.array([dev.var(O1.obs)])
-        expected = 0.25 * (3 - np.cos(2 * theta) - 2 * np.cos(theta) ** 2 * np.cos(2 * phi))
+        expected = 0.25 * (
+            3 - np.cos(2 * theta) - 2 * np.cos(theta) ** 2 * np.cos(2 * phi)
+        )
 
         self.assertAlmostEqual(var, expected, delta=3 / np.sqrt(shots))
 
-    def test_var_hermitian(self, shots, execution_timeout, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_var_hermitian(self, shots, execution_timeout, qvm, compiler, parallel):
         """Tests for variance calculation using an arbitrary Hermitian observable"""
-        dev = plf.QVMDevice(device="2q-qvm", shots=100 * shots, execution_timeout=execution_timeout)
+        dev = plf.QVMDevice(
+            device="2q-qvm",
+            shots=100 * shots,
+            execution_timeout=execution_timeout,
+            parallel=parallel,
+        )
 
         phi = 0.543
         theta = 0.6543
@@ -311,7 +344,9 @@ class TestQVMBasic(BaseTest):
     )
     def test_apply(self, op, apply_unitary, shots, qvm, compiler):
         """Test the application of gates to a state"""
-        dev = plf.QVMDevice(device="3q-qvm", shots=shots, parametric_compilation=False)
+        dev = plf.QVMDevice(
+            device="3q-qvm", shots=shots, parametric_compilation=False, parallel=True
+        )
 
         obs = qml.expval(qml.PauliZ(0))
 
@@ -344,11 +379,12 @@ class TestQVMBasic(BaseTest):
         # performing 1024 shots.
         self.assertAllAlmostEqual(res, expected, delta=3 / np.sqrt(shots))
 
-    def test_sample_values(self, qvm, tol):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_sample_values(self, qvm, tol, parallel):
         """Tests if the samples returned by sample have
         the correct values
         """
-        dev = plf.QVMDevice(device="1q-qvm", shots=10)
+        dev = plf.QVMDevice(device="1q-qvm", shots=10, parallel=parallel)
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(1.5708, wires=[0])
@@ -364,7 +400,8 @@ class TestQVMBasic(BaseTest):
         self.assertAllAlmostEqual(s1**2, 1, delta=tol)
         self.assertAllAlmostEqual(s1, 1 - 2 * dev._samples[:, 0], delta=tol)
 
-    def test_sample_values_hermitian(self, qvm, execution_timeout, tol):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_sample_values_hermitian(self, qvm, execution_timeout, tol, parallel):
         """Tests if the samples of a Hermitian observable returned by sample have
         the correct values
         """
@@ -372,7 +409,12 @@ class TestQVMBasic(BaseTest):
         shots = 1_000_000
         A = np.array([[1, 2j], [-2j, 0]])
 
-        dev = plf.QVMDevice(device="1q-qvm", shots=shots, execution_timeout=execution_timeout)
+        dev = plf.QVMDevice(
+            device="1q-qvm",
+            shots=shots,
+            execution_timeout=execution_timeout,
+            parallel=parallel,
+        )
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(theta, wires=[0])
@@ -396,10 +438,16 @@ class TestQVMBasic(BaseTest):
 
         # the analytic variance is 0.25*(sin(theta)-4*cos(theta))^2
         assert np.allclose(
-            np.var(s1), 0.25 * (np.sin(theta) - 4 * np.cos(theta)) ** 2, atol=0.1, rtol=0
+            np.var(s1),
+            0.25 * (np.sin(theta) - 4 * np.cos(theta)) ** 2,
+            atol=0.1,
+            rtol=0,
         )
 
-    def test_sample_values_hermitian_multi_qubit(self, qvm, execution_timeout, tol):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_sample_values_hermitian_multi_qubit(
+        self, qvm, execution_timeout, tol, parallel
+    ):
         """Tests if the samples of a multi-qubit Hermitian observable returned by sample have
         the correct values
         """
@@ -415,7 +463,12 @@ class TestQVMBasic(BaseTest):
             ]
         )
 
-        dev = plf.QVMDevice(device="2q-qvm", shots=shots, execution_timeout=execution_timeout)
+        dev = plf.QVMDevice(
+            device="2q-qvm",
+            shots=shots,
+            execution_timeout=execution_timeout,
+            parallel=parallel,
+        )
 
         with qml.tape.QuantumTape() as tape:
             qml.RX(theta, wires=[0])
@@ -465,21 +518,25 @@ class TestQVMBasic(BaseTest):
     def test_raise_error_if_shots_is_not_positive(self, shots):
         """Test that instantiating a QVMDevice if the number of shots is not a postivie
         integer raises an error"""
-        with pytest.raises(ValueError, match="Number of shots must be a positive integer."):
-            dev = plf.QVMDevice(device="2q-qvm", shots=shots)
+        with pytest.raises(
+            ValueError, match="Number of shots must be a positive integer."
+        ):
+            plf.QVMDevice(device="2q-qvm", shots=shots)
 
     def test_raise_error_if_shots_is_none(self, shots):
         """Test that instantiating a QVMDevice to be used for analytic computations raises an error"""
         with pytest.raises(
             ValueError, match="QVM device cannot be used for analytic computations."
         ):
-            dev = plf.QVMDevice(device="2q-qvm", shots=None)
+            plf.QVMDevice(device="2q-qvm", shots=None)
 
     @pytest.mark.parametrize("device", ["2q-qvm", np.random.choice(TEST_QPU_LATTICES)])
     def test_timeout_set_correctly(self, shots, device):
         """Test that the timeout attrbiute for the QuantumComputer stored by the QVMDevice
         is set correctly when passing a value as keyword argument"""
-        dev = plf.QVMDevice(device=device, shots=shots, compiler_timeout=100, execution_timeout=101)
+        dev = plf.QVMDevice(
+            device=device, shots=shots, compiler_timeout=100, execution_timeout=101
+        )
         assert dev.qc.compiler._timeout == 100
         assert dev.qc.qam._qvm_client.timeout == 101
 
@@ -508,8 +565,8 @@ class TestQVMBasic(BaseTest):
             qml.RX(theta, wires=[0])
             qml.RX(phi, wires=[1])
             qml.CNOT(wires=[0, 1])
-            O1 = qml.expval(qml.Identity(wires=[0]))
-            O2 = qml.expval(qml.Identity(wires=[1]))
+            qml.expval(qml.Identity(wires=[0]))
+            qml.expval(qml.Identity(wires=[1]))
 
         dev.apply(tape.operations, rotations=tape.diagonalizing_gates)
 
@@ -528,7 +585,7 @@ class TestQVMBasic(BaseTest):
         with qml.tape.QuantumTape() as tape:
             qml.RZ(theta, wires=[0])
             qml.CZ(wires=[0, 1])
-            O1 = qml.expval(qml.PauliZ(wires=[0]))
+            qml.expval(qml.PauliZ(wires=[0]))
 
         dev.apply(tape.operations, rotations=tape.diagonalizing_gates)
 
@@ -536,13 +593,25 @@ class TestQVMBasic(BaseTest):
 
         assert dev.compiled_program == compiled_program
 
+    def test_error_if_parallel_with_no_threads(self):
+        """Test that an error is raised if the parallel option is set to True but max_threads
+        is 0 or less."""
+        with pytest.raises(
+            ValueError,
+            match="max_threads must be set to a positive integer greater than 0",
+        ):
+            plf.QVMDevice(device="2q-qvm", shots=5, parallel=True, max_threads=0)
+
 
 class TestParametricCompilation(BaseTest):
     """Test that parametric compilation works fine and the same program only compiles once."""
 
-    def test_compiled_program_was_stored_in_dict(self, qvm, mock_qvm, monkeypatch):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_compiled_program_was_stored_in_dict(
+        self, qvm, mock_qvm, monkeypatch, parallel
+    ):
         """Test that QVM device stores the compiled program correctly in a dictionary"""
-        dev = qml.device("rigetti.qvm", device="2q-qvm")
+        dev = qml.device("rigetti.qvm", device="2q-qvm", parallel=parallel)
         theta = 0.432
         phi = 0.123
 
@@ -550,8 +619,8 @@ class TestParametricCompilation(BaseTest):
             qml.RX(theta, wires=[0])
             qml.RX(phi, wires=[1])
             qml.CNOT(wires=[0, 1])
-            O1 = qml.expval(qml.Identity(wires=[0]))
-            O2 = qml.expval(qml.Identity(wires=[1]))
+            qml.expval(qml.Identity(wires=[0]))
+            qml.expval(qml.Identity(wires=[1]))
 
         dev.apply(tape.operations, rotations=tape.diagonalizing_gates)
 
@@ -571,12 +640,18 @@ class TestParametricCompilation(BaseTest):
             assert len(dev._compiled_program_dict.items()) == 1
             assert mock_qvm.compile.call_count == 1
 
+    @pytest.mark.parametrize("parallel", [False, True])
     def test_parametric_compilation_with_numeric_and_symbolic_queue(
-        self, mock_qvm, execution_timeout
+        self, mock_qvm, execution_timeout, parallel
     ):
         """Tests that a program containing numeric and symbolic variables as
         well is only compiled once."""
-        dev = qml.device("rigetti.qvm", device="2q-qvm", execution_timeout=execution_timeout)
+        dev = qml.device(
+            "rigetti.qvm",
+            device="2q-qvm",
+            execution_timeout=execution_timeout,
+            parallel=parallel,
+        )
 
         param1 = np.array(1, requires_grad=False)
         param2 = np.array(2, requires_grad=True)
@@ -652,10 +727,15 @@ class TestQVMIntegration(BaseTest):
         with pytest.raises(TypeError, match="missing 1 required positional argument"):
             qml.device("rigetti.qvm")
 
-        with pytest.raises(ValueError, match="Number of shots must be a positive integer"):
+        with pytest.raises(
+            ValueError, match="Number of shots must be a positive integer"
+        ):
             qml.device("rigetti.qvm", "2q-qvm", shots=0)
 
-    def test_qubit_unitary(self, shots, compiler_timeout, execution_timeout, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_qubit_unitary(
+        self, shots, compiler_timeout, execution_timeout, qvm, compiler, parallel
+    ):
         """Test that an arbitrary unitary operation works"""
         dev1 = qml.device(
             "rigetti.qvm",
@@ -664,6 +744,7 @@ class TestQVMIntegration(BaseTest):
             compiler_timeout=compiler_timeout,
             execution_timeout=execution_timeout,
             parametric_compilation=False,
+            parallel=parallel,
         )
         dev2 = qml.device(
             "rigetti.qvm",
@@ -672,6 +753,7 @@ class TestQVMIntegration(BaseTest):
             compiler_timeout=compiler_timeout,
             execution_timeout=execution_timeout,
             parametric_compilation=False,
+            parallel=parallel,
         )
 
         def circuit():
@@ -697,13 +779,18 @@ class TestQVMIntegration(BaseTest):
     @flaky(max_runs=10, min_passes=2)
     @pytest.mark.parametrize("device", ["2q-qvm", np.random.choice(TEST_QPU_LATTICES)])
     @pytest.mark.parametrize("requires_grad", [True, False])
-    def test_one_qubit_wavefunction_circuit(self, device, qvm, compiler, requires_grad):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_one_qubit_wavefunction_circuit(
+        self, device, qvm, compiler, requires_grad, parallel
+    ):
         """Test that the wavefunction plugin provides correct result for simple circuit.
 
         As the results coming from the qvm are stochastic, a constraint of 2 out of 5 runs was added.
         """
         shots = 100_000
-        dev = qml.device("rigetti.qvm", device=device, shots=QVM_SHOTS)
+        dev = qml.device(
+            "rigetti.qvm", device=device, shots=QVM_SHOTS, parallel=parallel
+        )
 
         a = 0.543
         b = 0.123
@@ -717,16 +804,21 @@ class TestQVMIntegration(BaseTest):
             qml.Rot(x, y, z, wires=0)
             return qml.expval(qml.PauliZ(0))
 
-        self.assertAlmostEqual(circuit(a, b, c), np.cos(a) * np.sin(b), delta=3 / np.sqrt(shots))
+        self.assertAlmostEqual(
+            circuit(a, b, c), np.cos(a) * np.sin(b), delta=3 / np.sqrt(shots)
+        )
 
     @flaky(max_runs=10, min_passes=3)
     @pytest.mark.parametrize("device", ["2q-qvm", np.random.choice(TEST_QPU_LATTICES)])
-    def test_2q_gate(self, device, qvm, compiler):
+    @pytest.mark.parametrize("parallel", [False, True])
+    def test_2q_gate(self, device, qvm, compiler, parallel):
         """Test that the two qubit gate with the PauliZ observable works correctly.
 
         As the results coming from the qvm are stochastic, a constraint of 3 out of 5 runs was added.
         """
-        dev = qml.device("rigetti.qvm", device=device, shots=QVM_SHOTS)
+        dev = qml.device(
+            "rigetti.qvm", device=device, shots=QVM_SHOTS, parallel=parallel
+        )
 
         @qml.qnode(dev)
         def circuit():
@@ -866,7 +958,9 @@ class TestQVMIntegration(BaseTest):
 
         call_history = []
         with monkeypatch.context() as m:
-            m.setattr(QuantumComputer, "compile", lambda self, prog: call_history.append(prog))
+            m.setattr(
+                QuantumComputer, "compile", lambda self, prog: call_history.append(prog)
+            )
 
             for i in range(5):
                 circuit(params)
@@ -877,7 +971,9 @@ class TestQVMIntegration(BaseTest):
 
     @flaky(max_runs=10, min_passes=1)
     @pytest.mark.parametrize("device", ["2q-qvm", np.random.choice(TEST_QPU_LATTICES)])
-    def test_compiled_program_was_correct_compared_with_default_qubit(self, qvm, device, tol):
+    def test_compiled_program_was_correct_compared_with_default_qubit(
+        self, qvm, device, tol
+    ):
         """Test that QVM device stores the compiled program correctly by comparing it with default.qubit.
 
         As the results coming from the qvm are stochastic, a constraint of 1 out of 5 runs was added.
@@ -904,7 +1000,9 @@ class TestQVMIntegration(BaseTest):
 
     @flaky(max_runs=10, min_passes=3)
     @pytest.mark.parametrize("device", ["2q-qvm", np.random.choice(TEST_QPU_LATTICES)])
-    def test_2q_gate_pauliz_pauliz_tensor_parametric_compilation_off(self, device, qvm, compiler):
+    def test_2q_gate_pauliz_pauliz_tensor_parametric_compilation_off(
+        self, device, qvm, compiler
+    ):
         """Test that the PauliZ tensor PauliZ observable works correctly, when parametric compilation
         was turned off.